+++ internal/prometheus/conventions.go
@@ -1,7 +1,5 @@
 package prometheus
 
-import "time"
-
 const (
 	// Metrics.
 	sliErrorMetricFmt = "slo:sli_error:ratio_rate%s"
@@ -16,13 +14,4 @@ const (
 	sloModeLabelName      = "sloth_mode"
 	sloSpecLabelName      = "sloth_spec"
 	sloObjectiveLabelName = "sloth_objective"
-
-	// Time windows.
-	periodTimeWindow30d = 30 * 24 * time.Hour
-	periodTimeWindow28d = 28 * 24 * time.Hour
 )
-
-var SupportedTimeWindows = map[time.Duration]struct{}{
-	periodTimeWindow30d: {},
-	periodTimeWindow28d: {},
-}