+++ test/integration/prometheus/generate_test.go
@@ -96,8 +96,13 @@ func TestPrometheusGenerate(t *testing.T) {
 		},
 
 		"Generate using 28 day time window should generate Prometheus rules.": {
-			genCmdArgs: "--window-days 28 --input ./testdata/in-base.yaml",
+			genCmdArgs: "--default-period-windows 672h --input ./testdata/in-base.yaml",
 			expOut:     expectLoader.mustLoadExp("./testdata/out-base-28d.yaml.tpl"),
+		},
+
+		"Generate using custom 7 day time window should generate Prometheus rules.": {
+			genCmdArgs: "--default-period-windows 168h --input ./testdata/in-base.yaml --slo-period-windows-path ./windows",
+			expOut:     expectLoader.mustLoadExp("./testdata/out-base-custom-windows-7d.yaml.tpl"),
 		},
 
 		"Generate using invalid version should fail.": {